Moves scripts into own  package
